scripts/iok-converter/pyproject.toml
[project]
name = "iok-converter"
version = "0.1.0"
description = "Converts IOK (Indicator of Kit) rules to Sublime Security detection rules"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fire>=0.7.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
]

[project.scripts]
iok-converter = "iok_converter.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/iok_converter"] 
